Secured edit route
