+++ middleware.js
@@ -1,3 +1,8 @@
+const {campgroundSchema, reviewSchema} = require('./schema');
+const ExpressError =require('./utils/ExpressError');
+const Campground = require('./models/campgrounds');
+
+
 module.exports.isLoggedIn = (req,res,next)=>{
     if(!req.isAuthenticated()){
         req.session.returnTo = req.originalUrl;
@@ -7,3 +12,34 @@ module.exports.isLoggedIn = (req,res,next)=>{
     }
     next();
 }
+
+module.exports.isAuthor = async(req,res,next) =>{
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)){
+        req.flash('error', "You don't have permission.");
+        return res.redirect(`/campgrounds/${id}`);
+    }
+    next();
+}
+
+module.exports.validateCampground = (req,res,next) =>{
+    const {error} = campgroundSchema.validate(req.body)
+    if (error){
+        const msg = error.details.map(el => el.message).join(',')
+        throw new ExpressError(msg, 400)
+    }
+    else{
+        next();
+    }
+}
+
+module.exports.validateReview = (req, res, next) => {
+	const { error } = reviewSchema.validate(req.body);
+	if (error) {
+		const msg = error.details.map((el) => el.message).join(",");
+		throw new ExpressError(msg, 400);
+	} else {
+		next();
+	}
+};
