+++ routes/campgrounds.js
@@ -1,23 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const {campgroundSchema} = require('../schema');
 const catchAsync =require('../utils/catchAsync');
-const ExpressError =require('../utils/ExpressError');
 const Campground = require('../models/campgrounds');
-const {isLoggedIn} = require('../middleware');
-const { authenticate } = require("passport");
+const {isLoggedIn, isAuthor,validateCampground} = require('../middleware');
 
-
-const validateCampground = (req,res,next) =>{
-    const {error} = campgroundSchema.validate(req.body)
-    if (error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    }
-    else{
-        next();
-    }
-}
 
 router.get('/', catchAsync(async (req, res) => {
     const campgrounds = await Campground.find({});
@@ -33,10 +19,6 @@ router.post('/',isLoggedIn,validateCampground, catchAsync(async (req, res) => {
     req.flash('success','Successfully made a new Campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }))
-router.get('/:id/edit',isLoggedIn, catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id)
-    res.render('campgrounds/edit', { campground })
-}))
 
 router.get('/:id',catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id).populate('reviews').populate('author');
@@ -44,16 +26,27 @@ router.get('/:id',catchAsync(async (req, res) => {
         req.flash('error', "Campground Doesn't Exist!");
         return res.redirect('/campgrounds');
     }
+    
     res.render('campgrounds/show', { campground })
 }))
 
-router.put('/:id',isLoggedIn,validateCampground, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if (!campground){
+        req.flash('error', "Campground Doesn't Exist!");
+        return res.redirect('/campgrounds');
+    }
+    res.render('campgrounds/edit', { campground })
+}))
+
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
     const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Updated Successfully!');
     res.redirect(`/campgrounds/${campground._id}`);
 }))
-router.delete('/:id',isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id',isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Deleted Successfully!');
