+++ routes/reviews.js
@@ -1,21 +1,9 @@
 const express = require("express");
 const router = express.Router({mergeParams:true});
-const {reviewSchema} = require('../schema');
-const ExpressError =require('../utils/ExpressError');
+const { validateReview } = require('../middleware');
 const catchAsync =require('../utils/catchAsync');
 const Review = require('../models/review');
 const Campground = require('../models/campgrounds');
-
-
-const validateReview = (req, res, next) => {
-	const { error } = reviewSchema.validate(req.body);
-	if (error) {
-		const msg = error.details.map((el) => el.message).join(",");
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-};
 
 router.post('/',validateReview, catchAsync(async(req,res)=>{
     const campground = await Campground.findById(req.params.id);
